+++ CHANGELOG.md
@@ -34,6 +34,9 @@
 * SPDistributedCacheService
   * Improved error message for inclusion of server name into ServerProvisionOrder
     parameters when Present or change to Ensure Absent
+* SPSearchContentSource
+  * Add CrawlVirtualServers and CrawlSites CrawlSetting for SharePoint content
+    sources.  Fix for issue #1124.
 
 ## v3.6