+++ CHANGELOG.md
@@ -12,10 +12,13 @@ resources:
 * SPWebAppAuthentication
   * Updated resource to add support for Basic Authentication
 
-## UNRELEASED
+## v3.7
 
 * SPConfigWizard
   * Fixed issue with incorrect check for upgrade status of server
+* SPDistributedCacheService
+  * Improved error message for inclusion of server name into ServerProvisionOrder
+    parameters when Present or change to Ensure Absent
 * SPFarm
   * Removed SingleServer as ServerRole, since this is an invalid role.
   * Handle case where null or empty CentralAdministrationUrl is passed in